Added sqrt function to Matrix. Will be useful in localization routines.
Note: because of the name of the function, Matrix::sqrt and std::sqrt have to be distinguished inside the matrix class. But that should only be necessary inside matrix.cc! So if you want the sqrt of a number, you need to call ::sqrt(number). For now that seems fine. I can just change the name of the Matrix::sqrt() function if it is really annoying.