Added set_active function in Reference. It can take either a set<int> or a string of the form "(1, 2, 3)". If it takes the set, the MO numbers are used as is. If it takes the string, it subtracts one from each number (basically just assuming it came from the input and started counting from 1).
I went with this design because it was the only way to not have to redesign the const-ness of the entire class.

It'd also be nice to have some general function that can take a string of the form "(1, 2, 3)" and return whatever STL container desired, but that can come later.