Lots more organizing and simplifying. I think I've finally got a good structure.
Derived classes of the Quantization object compute the actual operations of a^dagger b and so on so that the form_gamma function is passed one of these Quantization objects and it does the rest. This should really really streamline writing all the coupling code itself, of which there is a lot. But I've started.