MEH is now a template. It works with CAS. This should make it pretty easy to get MEH to work with RAS and with future methods too (if any).
The structure does leave something to be desired. Since it's all a template, it all needs to be in the header. But to keep the files readable, I separated the headers into different files that are called in the main header. These secondary headers are given ".hpp" extensions just to make them look different than the other headers.

The better way to do this would be to spin off as many of those other files as possible into other helper classes. Or maybe have an MEH_base with all of the things that don't depend on template parameters, then derive the template version from that and then all specific MEH classes from there. These are all things I'll think about once RAS is done.