Converted class RysInt into a template
It now compiles and passes the test suite, although there's certainly some room for cleanup, and I'm sure more variables will need to be converted from double to DataType in order for the complex branch to work.  Among other things, I had to change allocate_data into a virtual function and redefine it in ERIBatch_base and NAIBatch_Base.  Otherwise, we ended up with a circular set of #include commands.