London dfHF seems to be working at zero magnetic field.
I think the largest remaining problem is that, if we have complex coefficients, the Fock matrix comes out as nonHermitian.  The problem seems to be in the various DF functions called
within Fock::two_electron_part_with_coeff.  The function diagonalize_real in this commit prevents us from getting a complex (unitary) matrix of eigenvalues when diagonalizing a real symmetric matrix.