Expanding and standardizing constants list.
For all fundamental constants, I used the CODATA 2010 recommended values.  This resulted in a slight change in the definition of the speed of light and angstrom unit conversion.  The speed of light change seems to have had no effect, but the length unit conversion messed up some test cases.  I decided to update the test cases so that, upon conversion to atomic units, they would yield the same geometry as before.