Adding a 1-electron magnetic interaction to RelHCore_London
The term added in this commit can be derived from the Dirac Hamiltonian and our definition of the small component spinors.  It cancels out to zero without a magnetic field, since p_x and p_y commute.
I'm pretty sure this term is needed, but something is clearly wrong in the results it gives.  For now, the extra term is computed in relhcore_london.cc and then thrown away to avoid breaking test cases.