Reworking input for magnetic field calculations
Now, it decides which basis to use automatically based on whether or not a magnetic field is specified.  The input parameter basis_type can be used to run zero-field calculations with the London orbital code, or to use Gaussian orbitals (with a common gauge origin) in the presence of a magnetic field.