London integral optimization
The idea is to use the standard 1-D interpolation whenever all T values are real (when magnetic field is zero or certain basis functions are on the same center).  Seems to work well for DF methods, but for some reason this changes the energy of the non-density-fitted test case in the 8th decimal place.