Fixing the basis set projector for RMB basis
…which was completely wrong when originally written.

The meat of this change was accentually included in a previous commit focusing on orthonormal guess.
I use MixedBasisArray<SmallInts1e_London<ComplexOverlapBatch>> to build the small component mixed-overlap matrix.
That's fine, but it added a new constructor for ComplexOverlapBatch.  Properly SmallInts1e_London should be a variadic template, but
storing optional arguments for use later on is not straightforward.