Adding helper functions to test the symmetry of a matrix
I'd prefer the implementation be in matrix_base, but it does not have transpose() etc. due to non-covariant return types due to shared_ptr.  Now check that this symmetry is approximately present when averaging out the errors, and I've replaced the check hard-coded in ZMatrix::diagonalize() with a call to one of the new functions.  It uses rms() rather than norm()/size(), but should give about the same error measurement…