Necessary to compile on Zinc
In coeff.cc, slice() is called with mend = mstart+1.  The compiler thus strips out the assert(mend >= mstart) check in optimization.  In the rare case where mstart+1 causes integer overflow, this would lead to different behavior in optimized and non-optimized machine code, so it triggers the -Wstrict-overflow flag.

In principle we should be able to run this check, but I don't know of any cases where t's actually been triggered, so for now I will remove it.  Not sure why I didn't get the same error when compiling locally.