Another revision to ZFCI initial guess
Some time ago I set it to check both alpha and beta when skipping determinants.  This made it possible to get the whole eigenspace, but leads to linear dependency in the guess
vectors if we feed them through the singlet spin-adaptation function.  Now we only check both alpha and beta when skipping spin_adapt; for fewer states we check open & closed,
which should reproduce the original behavior.