Splitting off init_kramers_coeff() into two functions for DHF and nonrel CASSCF wavefunctions
It had a lot of "if" statements and relatively little code used in both cases.  More importantly,
this is a step toward combining ZCASSCF::init_kramers_coeff() with ZFCI::kramers_zquat(), which
serves the same basic purpose.  I'm not sure that the version for nonrel CASSCF inputs is working
right, but at least for now will keep it.