ZFCI::kramers_zquat() and ZCASSCF::init_kramers_coeff() now both return a shared_ptr<ZMatrix>.
This is mainly to enable merging the two.  Previously ZFCI returned a Kramers<2,ZMatrix> which we can easily build from the returned matrix, while
ZCASSCF would edit coeff_ inline.

Note that this can likely be optimized if necessary.  At least for ZFCI, we Kramers-adapt the full coefficient matrix then, apparently, only use the active part of it.