Bug fix
For an odd number of electrons, the RelCoeff generated from Dirac would miscount the number of orbitals and accidentally delete the last virtual.  Interestingly, this error was silent in ZFCI as written.

This fix sets it to have one "active" orbital for an odd number of electrons…  debatable, but it keeps things sorted properly.