Bug fix - Moves the init_kramers_coefficient() call to ZHarrison constructor, *before* setting active orbitals.
With this change, the ZFCI energy is fully invariant with respect to order of active orbitals, at least in Kramers-restricted cases.

In the process of making this change, I had to move the geom_->relativistic() call from RelMOFile to ZHarrison.  We also now build RelHcore twice (in ZHarrison and RelMOFile), which is not good.
Also removed a header dependency