Minor
Makes nonrel_to_relcoeff() take in the matrix being converted as a parameter.

I wanted to remove nr_coeff_ as a member since it is only used upon construction (and then only with a CASSCF Reference), but the fact that we overwrite ref_ with a  RelReference makes the way of doing this…  ugly.  When we finally eliminate the dependence on RelReference, we should get nr_coeff_ from there rather than saving it.