Compute NAI in batches when the number of charges is very large, to keep memory costs down
I had to add a template specialization for Small1e<NAIBatch>::computebatch() and its GIAO analogue.  Currently the splitting of the Molecule is done in computebatch().  I don't really like that, but to do it in NAIBatch would require major refactoring, and I can't have it in the constructor of Hcore without storing new members that would hardly be used…  I don't see any major timing cost to this choice.