Bug fix - serialization with relativity and GIAO
I forgot to adapt the Shell serialization to allow for RMB, so reference saving failed when for relativistic calculations with magnetic fields.  (It gave a segmentation fault upon loading.)
Unfortunately to fix this we need access to the magnetic field and whether or not we are using common-origin, parameters I had not wanted to save in each Shell.
For now they are saved as members of Shell.

An alternative design would end the loading of Shells without handling relativistic parameters, then call Geometry::relativistic() when loading the Geometry…
I chose the above design to match how the field-free code works.