Moving the original pseudo spin Hamiltonian from ZHarrison to prop/pseudospin
Much cleanup is needed.  Some debug code is still present, and we currently pass many members of ZHarrison, as well as the method itself.  (I wanted to avoid passing the Method and use a function pointer to access ZHarrison::rdm1(i, j), but apparently it doesn't work that way for member functions.)