Bug fix for S=3/2
We need to adjust the phase of the pseudospin eigenvectors to preserve time-reversal symmetry.  With this change, we get a consistent set of linear equations, and can fully reconstruct the Hamiltonian and energy spectrum, but there appears to still be some freedom in D and E.

This fix is only for S = 3/2 systems; probably I will need to find a more general approach.