Updating constants.h to the 2014 CODATA constants (which were released in June)
The only test case to be affected was my lithiumlike Iodine atom w/ magnetic field.  The total energy on that one is larger than most, so it makes sense that it would be more sensitive.  I have adjusted the expected result and slightly loosened the check threshold for that single test.