Time-reversal symmetry matrix in pseudospin state basis
The original implementation was good for AO and MO basis but had the same bug as was corrected in my most recent commit.
This is a "special case" implementation good for up to 4 active electrons.
It should be possible to make a general (and more efficient) version using modified 1RDM code, but it is a little bit complicated by the complex conjugation operator.

The "testconj" bool and related code are just for verification.