Refactoring pseudo spin mapping to clean up axis definitions
Spin is quantized along the main g-axis by default, but we can specify otherwise in the input.  Now, it defines a new x and y axis to match the given z, and also we compute true spin matrices, etc., using those axis vectors

Also some cleanup of naming and stuff.