Adds a test case for zero-field splitting parameters (Woo hoo!)
The test case is an excited quartet (S=3/2) of the O2 anion, and takes about 4 seconds with optimized code.  The energy differences are small, though; it might be better to replace this with a complex with measurable rhombic splitting (E > 0).  We also don't have a test case for electronic g-factors.