Bug fix
We use a diagonalization in forming kramers-adapted orbitals.  We need to use the orbitals that were occupied in the Hartree--Fock step that generated the input orbitals; if the
charge is changed this drops or adds some, which can break the Kramers-symmetry of the Fock matrix we then feed to zquatev for diagonalization.