For (X)MS-CASPT2 the Dyson orbitals should be calculated from the XMS-CASPT2 configuration interaction vectors, which are obtained by rotating the CASSCF vectors with the XMS-CASPT2 rotation matrix U,
  C_{XMS-CASPT} = C_{CASSCF} U

as explained in section 2.1.1 of Pitesa et al. (2021)
https://doi.org/10.1021/acs.jctc.1c00396

Currently the unrotated CASSCF reference with the XMS-CASPT2 energies
are saved. The resulting Dyson orbitals are in the wrong order and
the printed ionization energies and Dyson norms do not match.

This commit fixes this issue. A 'save_ref' block after a 'smith'
block saves the references with the rotated CI vectors.

I need to check if this breaks other functionality, such as XMS-CASPT gradients.